% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_modularity.R
\name{assess_modularity}
\alias{assess_modularity}
\title{Assess modularity}
\usage{
assess_modularity(
  g_CC_LCC = NULL,
  commMethods = c("fastgreedy", "multilev"),
  BPPARAM = NULL
)
}
\arguments{
\item{g_CC_LCC}{list of CC and LCC as produced by means of calc_CC_LCC()}

\item{commMethods}{community method(s) that will be considered}

\item{BPPARAM}{BBPARAM}
}
\description{
Assess modularity
}
