% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_communities.R
\name{assign_communities}
\alias{assign_communities}
\title{Assign communities}
\usage{
assign_communities(g = NULL, commList = NULL)
}
\arguments{
\item{g}{igraph object}

\item{commList}{community structure}
}
\description{
the community will be added as the vertex attribute "comm_id"
}
