% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CC_LCC.R
\name{calc_CC_LCC}
\alias{calc_CC_LCC}
\title{Calculate connected components and largest connected componets}
\usage{
calc_CC_LCC(G = NULL, vertices = NULL, ranks = NULL, vertices_max = 500)
}
\arguments{
\item{G}{igraph object}

\item{vertices}{set of (top ranking) vertices to study}

\item{ranks}{optional ranks at which the modularity will be computes}

\item{vertices_max}{maximum number of vertices allowed; the higher vertices_max the higher the computational cost}
}
\description{
Calculate connected components and largest connected componets
}
