% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gs_sim.R
\name{calc_gs_sim}
\alias{calc_gs_sim}
\title{Select the most representative features}
\usage{
calc_gs_sim(gsl = NULL, method = "Simpson")
}
\arguments{
\item{gsl}{gene set list}

\item{method}{on of the method available for proxy::simil()}
}
\description{
Select the most representative features
}
