% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_top_net_scores.R
\name{cmp_top_net_scores}
\alias{cmp_top_net_scores}
\title{Compare top networks scores}
\usage{
cmp_top_net_scores(NRRes = NULL, netEnrRes = NULL, top = 5, out_dir = NULL)
}
\arguments{
\item{NRRes}{dataframe resulting from network resampling from package dmfind `dmfind::NR()`}

\item{netEnrRes}{dataframe resulting from `assess_enrichment()`}

\item{top}{number of top ranks that will be plotted in red}

\item{out_dir}{output directory}
}
\description{
Define significance for enriched genes
}
