% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFDR.R
\name{eFDR}
\alias{eFDR}
\title{empirical False Discovery Rate}
\usage{
eFDR(real_values = NULL, all_values = NULL, BPPARAM = NULL, correct.max = TRUE)
}
\arguments{
\item{real_values}{vactor of real (observed) values}

\item{all_values}{vactor of all values (real + permuted) values}

\item{BPPARAM}{An optional BiocParallelParam instance determining the parallel back-end to be used during evaluation. If NULL, parallel evaluation is disabled using SerialParam(). See ?bplapply.}

\item{correct.max}{logical, to decide if to correct the maximum values to 1 or not}
}
\description{
empirical False Discovery Rate
}
