% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_sets_to_occ.R
\name{feat_sets_to_occ}
\alias{feat_sets_to_occ}
\title{Feature set list to occurrence matrix}
\usage{
feat_sets_to_occ(x = NULL)
}
\arguments{
\item{x}{gene set list}
}
\description{
Feature set list to occurrence matrix
}
