% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_cartography.R
\name{functional_cartography}
\alias{functional_cartography}
\title{pc_wmz Calculate participation coefficient and within module degree z-score}
\usage{
functional_cartography(
  topNetwork = NULL,
  labelBy = NULL,
  useP = FALSE,
  hCut = NULL
)
}
\arguments{
\item{topNetwork}{network with some parameters: "comm_id" for community
membership and "label" for labels (for plotting purpose).}

\item{labelBy}{vertex attribute of topNetwork to use for labels}

\item{useP}{whether to use p or not in the vertical axis}

\item{hCut}{value to divide the horizontal axis}
}
\value{
dataframe with pc, wmz, region, description and role
}
\description{
Calculate participation coefficient
and within module degree z-score
}
