% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets_to_X0.R
\name{gene_sets_to_X0}
\alias{gene_sets_to_X0}
\title{Quantify the occurence of gene sets among positive element of X0}
\usage{
gene_sets_to_X0(X0 = NULL, gsl = NULL, add_names = F)
}
\arguments{
\item{X0}{X0 matrix}

\item{gsl}{gene set list. The element must match the row names of X0}

\item{add_names}{TRUE/FALSE, whether to add a string with collapsed rown names of X0; if not only numbers will be included in the output}
}
\description{
Quantify the occurence of gene sets among positive element of X0
}
