% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdf.R
\name{get_cdf}
\alias{get_cdf}
\title{Gene-level cdf estimation}
\usage{
get_cdf(gene_expr = NULL, type = c("ecdf", "kcdf"))
}
\arguments{
\item{gene_expr}{normalized gene expression profile}

\item{type}{"kcdf" will use Gaussian kernel while "ecdf" will use empirical estimate}
}
\description{
Gene-level cdf estimation
}
