% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_genes.R
\name{get_cluster_genes}
\alias{get_cluster_genes}
\title{Get pathway clusters from enrichment map}
\usage{
get_cluster_genes(gene_set_clusters = NULL, gene_set_list = NULL)
}
\arguments{
\item{gene_set_clusters}{gene set cluster data frame created by get_gene_set_clusters()}

\item{gene_set_list}{gene set list}
}
\description{
Get pathway clusters from enrichment map
}
