% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_set_clusters.R
\name{get_gene_set_clusters}
\alias{get_gene_set_clusters}
\title{Get pathway clusters from enrichment map}
\usage{
get_gene_set_clusters(enrichment_map_graph = NULL)
}
\arguments{
\item{enrichment_map_graph}{igraph object created by enrichment_map()}
}
\description{
Get pathway clusters from enrichment map
}
