% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gs_score.R
\name{get_gs_score}
\alias{get_gs_score}
\title{Get a gene set score from Xs matrix}
\usage{
get_gs_score(Xs = NULL, gsl = NULL, FUN = NULL)
}
\arguments{
\item{Xs}{Xs matrix}

\item{gsl}{named list of gene sets, which must match the row names of X}

\item{FUN}{function to use to calculate the gene set score from the element of Xs; mean() by default}
}
\description{
Get a gene set score from Xs matrix
}
