% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kegg_db.R
\name{get_kegg_db}
\alias{get_kegg_db}
\title{Get KEGG Pathways using KEGGREST}
\usage{
get_kegg_db(org = "hsa", mapping = NULL, out_file = "kegg_pathway.rds")
}
\arguments{
\item{org}{organism, following KEGG terminology}

\item{mapping}{data.frame with two columns: NCBI Entrez gene_id and symbol}

\item{out_file}{output file name}
}
\description{
Get KEGG Pathways using KEGGREST
}
