% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND.R
\name{ND}
\alias{ND}
\title{Network Diffusion}
\usage{
ND(X0, W, alpha=0.7, nMax=1e4, eps=1e-6, finalSmooth=FALSE, 
fullOutput=FALSE, verbose=FALSE)
}
\arguments{
\item{X0}{vector or matrix composed of column vectors with initial 
distribution of information}

\item{W}{symmetrically normalized adjacency matrix W = D^-1 A D^-1, 
see normalize_adj_mat function}

\item{alpha}{numeric, the smothing factor}

\item{nMax}{numeric, maximum number of iterations}

\item{eps}{numeric, the iteration will stop when the maximum difference 
between matrix Xs between two consecutive iteraction is 
smaller than \code{eps}}

\item{finalSmooth}{TRUE/FALSE, whether to do the final step of smoothing}

\item{fullOutput, }{TRUE/FALSE, whether to output all steps}

\item{verbose, }{TRUE/FALSE}
}
\value{
A matrix with steady state values. If fullOutput is TRUE, a list with:
\item{\code{Xs}}{ the smoothed matrix;}
\item{\code{eps}}{ see above;}
\item{\code{maxAbsDiff}}{ maximum absolute difference between F(t) and F(t+1);}
\item{\code{XsAll}}{ transient Xs matrices.}
}
\description{
Network Diffusion
}
\examples{
\dontrun{ND(X0, W, alpha=0.7, nMax=1e4, eps=1e-6, finalSmooth=FALSE, 
fullOutput=FALSE, verbose=FALSE)}
}
