% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_nd.R
\name{normalize_nd}
\alias{normalize_nd}
\title{Normalization of ND steady state values}
\usage{
normalize_nd(x = NULL, method = c("max", "min_max", "sum"), ymax = 1, ymin = 0)
}
\arguments{
\item{x}{numeric vector with steady state diffusion values}

\item{method}{normalization approach}

\item{ymax}{max value for min-max normalization}

\item{ymin}{min value for min-max normalization}
}
\value{
Normalized ND values
}
\description{
Normalization of ND steady state values
}
