% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_mat.R
\name{perm_mat}
\alias{perm_mat}
\title{Permutation of X}
\usage{
perm_mat(X = NULL, perms = NULL, columns = FALSE)
}
\arguments{
\item{X}{matrix; a score matrix X, in which each column (layer) is a score vector over all vertices of A.}

\item{perms}{list of permutations of the rownames of X, obtained by means of \code{permute_vertices()}.}

\item{columns}{whether to apply the permutation to the columns.}
}
\value{
list with permutations of X (where the first element of the list contains real data)
}
\description{
Permutation of input matrix X
}
