% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_vertices.R
\name{perm_vertices}
\alias{perm_vertices}
\title{Permutation of graph vertices}
\usage{
perm_vertices(
  vert_deg = NULL,
  k = 99,
  seed_n = NULL,
  vertex_sets = NULL,
  method = c("simple", "degree"),
  cut_par = 20,
  bin_type = c("interval", "number", "width")
)
}
\arguments{
\item{vert_deg}{named vector of vertex degree}

\item{k}{number of permutations.}

\item{seed_n}{optional numeric; to specify the seed for (pseudo) random number generation; using the same seed}

\item{vertex_sets}{optional list with subsets of names(vert_deg); each subset will be permuteted separately; this is useful when distinct sets of vertices have different meaning;}

\item{method}{method to perform permutation; "simple" permutes vertices only based on vertex sets; "degree" permutes vertices based on vertex_sets and vertex degree}

\item{cut_par}{number of intervals or width of the intervals}

\item{bin_type}{"interval" will use cut_interval(), "number" will use cut_number() while "width" will use cut_width() function from ggplot2}
}
\value{
list with k+1 vectors of permutations of names(vert_deg), where the first element of the list contains original names(vert_deg)
}
\description{
Permutation of graph vertices
}
