% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fc.R
\name{plot_fc}
\alias{plot_fc}
\title{Plot functional cartography}
\usage{
plot_fc(
  topNetworks = NULL,
  labelBy = NULL,
  hCut = NULL,
  useP = FALSE,
  cex.pt = 0.8,
  cex.pt.lab = 0.8,
  font = 1,
  ...
)
}
\arguments{
\item{topNetworks}{igraph object}

\item{labelBy}{one of the vertex attributes}

\item{hCut}{where to place the division of the vertical axis}

\item{useP}{whether p is used in the vertical axis}

\item{cex.pt}{point size}

\item{cex.pt.lab}{point label size}

\item{font}{point label font type}

\item{...}{further arguments to plot()}
}
\description{
Plot functional cartography
}
