% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot a network}
\usage{
plot_network(
  graph = NULL,
  colorBy = NULL,
  colorQuant = TRUE,
  labelBy = "name",
  pal = NULL,
  community = NULL,
  commWin = 2,
  commWb = 1,
  lo = NULL,
  legendOff = FALSE,
  labelComm = TRUE,
  labelCommCex = 0.8,
  labelCommCol = "black",
  ...
)
}
\arguments{
\item{graph}{igraph object}

\item{colorBy}{object to be colored by an attribute}

\item{colorQuant}{TRUE/FALSE}

\item{labelBy}{select a feature to label the graph}

\item{pal}{palette to color the graph}

\item{community}{communities from network analysis}

\item{commWin}{distance of the vertices inside a community}

\item{commWb}{distance between communities}

\item{lo}{layout}

\item{legendOff}{legend}

\item{...}{additional arguments to `plot.igraph()`}
}
\description{
Extraction and plotting of connected subnetworks 
from a graph and a set of selected vertices
}
