% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_scores.R
\name{plot_network_scores}
\alias{plot_network_scores}
\title{Plot the network scoring summary}
\usage{
plot_network_scores(net_scores_summm = NULL)
}
\arguments{
\item{net_scores_summm}{the table provided by the function score_networks_summary()}
}
\value{
A ggplot2 object
}
\description{
Plot the values obtained by means of score_networks_summary()
}
