% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_networks.R
\name{score_networks}
\alias{score_networks}
\title{Score the networks composed of a ranked list of genes}
\usage{
score_networks(
  g_CC_LCC = NULL,
  X0 = NULL,
  X0_column = 1,
  NRsummary = NULL,
  commMethods = "fastgreedy",
  BPPARAM = NULL
)
}
\arguments{
\item{g_CC_LCC}{list of CC and LCC as produced by means of calc_CC_LCC()}

\item{X0}{the X0 matrix}

\item{X0_column}{the column of X0 that will be considered}

\item{NRsummary}{NRsummary element of the output of function NR()}

\item{commMethods}{community method(s) that will be considered}

\item{BPPARAM}{optional BBPARAM}
}
\value{
A data frame with the following quantities for CC and LCC:
\itemize{
\item \code{X0_}: sum of the X0 values in the network divided by the sum of the n genes with the highest X0 values, where n network size (CC or LCC);
\item \code{linkers}: fraction of linkers, that is, genes with X0(i) equal to 0;
\item \code{omega}: NR omega value;
\item \code{z}: omega z-score;
\item \code{p_val}: omega p_val;
\item \code{n_V_}: number of vertices;
\item \code{n_E_}: number of links;
\item \code{a_}: algorithm used to find the communities;
\item \code{Q_}: modularity;
\item \code{n_comm_}: number of communities.
}
}
\description{
The networks are scored by:
\itemize{
\item modularity;
\item network components;
\item omega z-score peaks;
\item presence of linkers.
}
}
