% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_networks_summary.R
\name{score_networks_summary}
\alias{score_networks_summary}
\title{Provide a summary score for he networks composed of a ranked list of genes}
\usage{
score_networks_summary(
  net_scores = NULL,
  LCC = FALSE,
  minpeakdistance = 1,
  minpeakheight = 0,
  ...
)
}
\arguments{
\item{net_scores}{the table provided by the function score_networks()}

\item{LCC}{whether to consider the largest connected components (TRUE) instead of considering all connected components (FALSE, default)}

\item{minpeakdistance}{the minimum distance (in indices) peaks have to have to be counted}

\item{minpeakheight}{the minimum (absolute) height a peak has to have to be recognized as such}

\item{...}{further argumnts for pracma::findpeaks()}
}
\value{
A data frame with:
\itemize{
\item the input scores normalized to their maximum; \code{not_linkers} is 1 minus linkers (divided to their maximum);
\item \code{y}: the summary score, which is the sum of the previous columns;
\item \code{peak}: the presence of a peak in y.
}
}
\description{
Provide a summary score for he networks composed of a ranked list of genes
}
