% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_modularity.R
\name{assess_modularity}
\alias{assess_modularity}
\title{Assess modularity}
\usage{
assess_modularity(
  G = NULL,
  vertices = NULL,
  ranks = NULL,
  commMethods = c("fastgreedy", "multilev"),
  BPPARAM = NULL,
  minComponentSize = 2
)
}
\arguments{
\item{G}{igraph object}

\item{vertices}{set of (top ranking) vertices to study}

\item{ranks}{ranks at which the modularity will be computes}

\item{commMethods}{community method(s) that will be considered}

\item{BPPARAM}{BBPARAM}

\item{minComponentSize}{minimum size of connected components}
}
\description{
Assess modularity
}
