% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_top_net_scores.R
\name{cmp_top_net_scores}
\alias{cmp_top_net_scores}
\title{Compare top networks scores}
\usage{
cmp_top_net_scores(
  NRRes = NULL,
  netEnrRes = NULL,
  column = 1,
  sigStatNR = "p",
  sigStatEn = "p",
  norm = FALSE,
  top = 5
)
}
\arguments{
\item{NRRes}{dataframe resulting from network resampling from package dmfind `dmfind::NR()`}

\item{netEnrRes}{dataframe resulting from `assess_enrichment()`}

\item{sigStatNR}{which p-value will be considered from NRRes}

\item{sigStatEn}{which p-value will be considered from netEnrRes}

\item{norm}{if TRUE (default) the two disitribution of -log10(p) will be mapped in the interval [1, 2]}

\item{top}{number of top ranks that will be plotted in red}

\item{columm}{which column to consider; in case multiple runs have been done}
}
\description{
Define significance for enriched genes
}
