% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_map.R
\name{enrichment_map}
\alias{enrichment_map}
\title{Enrichment map}
\usage{
enrichment_map(
  gs_scores = NULL,
  gene_set_sim = NULL,
  min_sim = 0.7,
  clustering_f = cluster_fast_greedy,
  min_comp_size = 1,
  gs_size = NULL
)
}
\arguments{
\item{gs_scores}{named vector of pathway scores}

\item{gene_set_sim}{gene set similarity calculated through `calc_gs_sim()`}

\item{min_sim}{threshold for the similarity score between two gene sets}

\item{clustering_f}{greph clustering function, `igraph::cluster_fast_greedy()` by default}

\item{min_comp_size}{minimum size of a community to be considered in the enrichment map
then all gene-sets are displayed}

\item{gs_size}{named vector with size of the gene-sets in `gs_scores`}
}
\value{
The function returns a list containing: 
\itemize{
 \item igraph = network object used for plotting the enrichment map
 \item network_data = data.frame with layout (X1,X2 columns), name of the gene-sets together with community ids ("comm_id"), 
  score and number of genes ("n_genes")
 \item path_comm_genes = list composed by the genes present in each community
 \item sim_coeff = data.frame with the similarity score calculated between each gene-set pair
 \item plot = enrichment map plot obtained by using `ggraph` package functions. Only if `save_plot = FALSE`
}
}
\description{
This functions calculates similarities between gene-sets and plot a resulting enrichment map
}
\details{
enrichment_map() function calculates similarities between each gene-set pair by using `method` metric. Subsequently,
 these similarities are filtered to maintain the ones >= `min_sim`. `clustering_f` function is then used to identify communities, 
 which may be filtered to plot only the ones composed by at least `min_comp_size` gene-sets.
}
