% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_X0.R
\name{perm_X0}
\alias{perm_X0}
\title{Permutation of X0}
\usage{
perm_X0(X0 = NULL, perms = NULL)
}
\arguments{
\item{X0}{matrix; a score matrix X0, in which each column (layer) is a score vector over all vertices of A.}

\item{perms}{list of permutations of the rownames of X0, obtained by means of `permute_vertices()`}
}
\value{
list with permutations of X0 (where the first element of the list contains real data)
}
\description{
Permutation of input matrix X0
}
