% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fc.R
\name{plot_fc}
\alias{plot_fc}
\title{Plot the functional cartography of the network}
\usage{
plot_fc(topNetworks = NULL, labelBy = NULL, hCut = NULL, useP = FALSE)
}
\arguments{
\item{topNetworks}{igraph object}

\item{labelBy}{optional, one of the vertex attributes}

\item{hCut}{optional, where to place the division of the vertical axis}

\item{useP}{whether the p-value associated with the z-score was used}
}
\value{
ggplo2 object
}
\description{
Plot the functional cartography of the network
}
