% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fc.R
\name{plot_fc}
\alias{plot_fc}
\title{Plot functional cartography}
\usage{
plot_fc(
  topNetworks = NULL,
  labelBy = NULL,
  hCut = NULL,
  useP = FALSE,
  out_dir = NULL
)
}
\arguments{
\item{topNetworks}{igraph object}

\item{labelBy}{one of the vertex attributes}

\item{hCut}{where to place the division of the vertical axis}

\item{useP}{whether p is used in the vertical axis}

\item{out_dir}{output directory}
}
\description{
Plot functional cartography
}
