% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mod_trend.R
\name{plot_modu_trend}
\alias{plot_modu_trend}
\title{Plot modularity trend}
\usage{
plot_modu_trend(assessModRes = NULL)
}
\arguments{
\item{assessModRes}{result of `assess_modularity()`}
}
\value{
plot
}
\description{
Plot modularity trend
}
