% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mod_trend.R
\name{plot_modu_trend}
\alias{plot_modu_trend}
\title{Plot modularity trend}
\usage{
plot_S(nsiRes=NULL, X0, initSigGenes=NULL, file="S.jpg")
}
\arguments{
\item{assessModRes}{result of `assess_modularity()`}

\item{file}{filename}
}
\value{
plot
}
\description{
Plot modularity trend
}
\examples{
\dontrun{plot_S(nsiRes=NULL, X0, initSigGenes=NULL, file="S.jpg")}
}
